# api/openapi/courses-organization.yaml
openapi: 3.0.3
info:
  title: Authena course-organization service
  description: service for course managing
  version: 0.0.1
servers:
  - url: 'https://{hostname}/v1'
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: [ ]
paths:
...
  /courses/{courseId}/tasks:
    get:
      tags:
        - tasks
      operationId: getCourseTasks
      description: returns list of course tasks
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/TaskType'
          required: false
          description: type of task for filtering
        - in: query
          name: title
          schema:
            type: string
          required: false
          description: task title substring for filtering
        - in: query
          name: description
          schema:
            type: string
          required: false
          description: task description substring for filtering
      responses:
        '200':
          description: found tasks of course
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - tasks
      operationId: addTaskToCourse
      description: creates task
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/CreateAutoCodeCheckingTaskRequest'
                - $ref: '#/components/schemas/CreateTestingTaskRequest'
                - $ref: '#/components/schemas/CreateManualCheckingTaskRequest'
              discriminator:
                propertyName: type
      responses:
        '201':
          description: manual checking task created
          headers:
            Content-Location:
              description: created task url
              schema:
                type: string
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: only authorized teacher can create task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}/tasks/{taskNumber}:
    get:
      tags:
        - tasks
      operationId: getCourseTask
      description: returns task of course
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
        - in: path
          name: taskNumber
          schema:
            type: integer
            minimum: 1
          required: true
          description: task number
      responses:
        '200':
          description: found task in course
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ManualCheckingTaskResponse'
                  - $ref: '#/components/schemas/AutoCodeCheckingTaskResponse'
                  - $ref: '#/components/schemas/TestingTaskResponse'
                discriminator:
                  propertyName: type
        '404':
          description: course or task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
    ManualCheckingTaskPart:
      type: object
      properties:
        deadline:
          $ref: '#/components/schemas/Deadline'
    AutoCodeCheckingTaskPart:
      type: object
      properties:
        testData:
          type: array
          items:
            $ref: '#/components/schemas/TestData'
        deadline:
          $ref: '#/components/schemas/Deadline'
    TestingTaskPart:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/TestPoint'
    CreateTaskRequest:
      $ref: '#/components/schemas/Task'
    TaskResponse:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            number:
              type: integer
              minimum: 1
    CreateManualCheckingTaskRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTaskRequest'
        - $ref: '#/components/schemas/ManualCheckingTaskResponse'
    CreateAutoCodeCheckingTaskRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTaskRequest'
        - $ref: '#/components/schemas/AutoCodeCheckingTaskResponse'
    CreateTestingTaskRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTaskRequest'
        - $ref: '#/components/schemas/TestingTaskPart'
    ManualCheckingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/ManualCheckingTaskPart'
    AutoCodeCheckingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/AutoCodeCheckingTaskPart'
    TestingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/TestingTaskPart'
...
    TestPoint:
      type: object
      required: [ description, variants, correctVariantNumbers ]
      properties:
        description:
          type: string
        variants:
          type: array
          items:
            type: string
        correctVariantNumbers:
          type: array
          items:
            type: integer
            minimum: 0
    TestData:
      type: object
      required: [ inputData, outputData ]
      properties:
        inputData:
          type: string
        outputData:
          type: string
    Deadline:
      type: object
      required: [ goodGradeTime, excellentGradeTime ]
      properties:
        goodGradeTime:
          type: string
          format: date
        excellentGradeTime:
          type: string
          format: date
...
    TaskType:
      type: string
      enum:
        - MANUAL_CHECKING
        - AUTO_CODE_CHECKING
        - TESTING
...
    Error:
      type: object
      required: [ slug, details ]
      properties:
        slug:
          type: string
        details:
          type: string